requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
cryptography==39.0.0
cycler==0.11.0
fonttools==4.38.0
frozendict==2.3.4
html5lib==1.1
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.2
multitasking==0.0.11
numpy==1.23.3
packaging==22.0
pandas==1.4.4
Pillow==9.4.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.13
webencodings==0.5.1
yahoofinancials==1.6
yfinance==0.2.3
